+++ src/main/java/org/itxtech/synapseapi/multiprotocol/protocol112/protocol/ClientCacheStatusPacket112.java
@@ -3,7 +3,7 @@ package org.itxtech.synapseapi.multiprotocol.protocol112.protocol;
 import cn.nukkit.network.protocol.ProtocolInfo;
 
 public class ClientCacheStatusPacket112 extends Packet112 {
-    public static final byte NETWORK_ID = ProtocolInfo.CLIENT_CACHE_STATUS_PACKET;
+    public static final int NETWORK_ID = ProtocolInfo.CLIENT_CACHE_STATUS_PACKET;
 
     public boolean supported;
